Refactored address format. Created ASP Django app
