+++ config/settings/base.py
@@ -69,6 +69,7 @@ LOCAL_APPS = [
     "itou.invitations",
     "itou.external_data.apps.ExternalDataConfig",
     "itou.metabase",
+    "itou.asp",
     # www.
     "itou.www.apply",
     "itou.www.approvals_views",
