+++ itou/asp/admin.py
@@ -1,0 +1,4 @@
+from django.contrib import admin
+
+
+# Register your models here.
