+++ itou/asp/tests.py
@@ -1,0 +1,4 @@
+from django.test import TestCase
+
+
+# Create your tests here.
